pyproject.toml
[project]
name = "hello-gpt-oss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai==1.99.1",
    "openai-agents==0.2.4",
    "python-dotenv",
]
